Extract chat app to common folder
