+++ packages/react-composites/package.json
@@ -18,7 +18,7 @@
     "tsc:e2e": "tsc -project tests/tsconfig.json",
     "build:e2e": "rushx _by-flavor \"rushx build:e2e:call && rushx build:e2e:chat && rushx build:e2e:callwithchat\"",
     "build:e2e:call": "rushx _by-flavor \"webpack-cli build --config ./tests/app/call/webpack.config.js --mode=production --env production\"",
-    "build:e2e:chat": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/chat/app/webpack.config.js --mode=production --env production\"",
+    "build:e2e:chat": "rushx _by-flavor \"webpack-cli build --config ./tests/app/chat/webpack.config.js --mode=production --env production\"",
     "build:e2e:callwithchat": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/callwithchat/app/webpack.config.js --mode=production --env production\"",
     "test": "rushx _by-flavor \"rushx _test:by-flavor\"",
     "test:coverage": "rushx _by-flavor \"rushx test:_by-flavor -- --coverage\"",
