+++ packages/react-composites/tests/app/chat/CustomDataModel.tsx
@@ -5,7 +5,7 @@ import { ChatParticipant } from '@azure/communication-signaling';
 import { Stack } from '@fluentui/react';
 import { MessageProps } from '@internal/react-components';
 import React from 'react';
-import { AvatarPersonaData } from '../../../../src';
+import { AvatarPersonaData } from '../../../src';
 
 /**
  * Custom onRenderTypingIndicator prop of {@link ChatComposite} used for UI test
