+++ packages/react-composites/tests/app/chat/FakeChatAdapterArgs.ts
