+++ packages/react-composites/tests/app/chat/webpack.config.js
@@ -3,8 +3,8 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
 module.exports = () => {
-  const babelConfig = require('../../../../.babelrc.js');
-  const commonConfig = require('../../common/testapp.webpack.config')(__dirname, babelConfig);
+  const babelConfig = require('../../../.babelrc.js');
+  const commonConfig = require('../testapp.webpack.config')(__dirname, babelConfig);
 
   return commonConfig;
 };
